testing/web-platform/tests/workers/WorkerGlobalScope_XMLHttpRequest.htm

<!DOCTYPE html>
<title> WorkerGlobalScope XMLHttpRequest() support </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function(t) {
  var worker = new Worker('./support/XMLHttpRequest.js');
  worker.onmessage = t.step_func_done(function(e) {
    assert_equals(e.data, "Request Succeeded!");
  });
  worker.postMessage('XMLHttpRequest.txt');
});
</script>
